# backend/docs/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrainerPlatform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/certification-core/competency-frameworks": {
      "get": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "List Frameworks",
        "description": "List competency frameworks with optional filters.",
        "operationId": "list_frameworks_api_v1_certification_core_competency_frameworks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "domain_pack_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Domain Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyFrameworkListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Create Framework",
        "description": "Create a new competency framework with competencies.",
        "operationId": "create_framework_api_v1_certification_core_competency_frameworks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyFrameworkCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyFrameworkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/certification-core/competency-frameworks/{framework_id}": {
      "get": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Get Framework",
        "description": "Get a competency framework by its business ID.",
        "operationId": "get_framework_api_v1_certification_core_competency_frameworks__framework_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "framework_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Framework Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyFrameworkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Update Framework",
        "description": "Update a competency framework.",
        "operationId": "update_framework_api_v1_certification_core_competency_frameworks__framework_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "framework_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Framework Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyFrameworkUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyFrameworkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/certification-core/competency-frameworks/{framework_id}/competencies": {
      "post": {
        "tags": [
          "Certification-Core",
          "Certification-Core"
        ],
        "summary": "Add Competency",
        "description": "Add a competency to a framework.",
        "operationId": "add_competency_api_v1_certification_core_competency_frameworks__framework_id__competencies_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "framework_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Framework Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetencyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompetencyCreate": {
        "properties": {
          "competency_id": {
            "type": "string",
            "maxLength": 100,
            "title": "Competency Id",
            "description": "Unique ID within framework"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "cognitive_levels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cognitive Levels"
          },
          "critical": {
            "type": "boolean",
            "title": "Critical",
            "default": false
          },
          "weight": {
            "type": "number",
            "title": "Weight",
            "default": 0.0
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "competency_id",
          "name"
        ],
        "title": "CompetencyCreate"
      },
      "CompetencyFrameworkCreate": {
        "properties": {
          "framework_id": {
            "type": "string",
            "maxLength": 100,
            "title": "Framework Id"
          },
          "domain_pack_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Pack Id"
          },
          "version": {
            "type": "string",
            "maxLength": 20,
            "title": "Version"
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "default": "en-US"
          },
          "market": {
            "type": "string",
            "title": "Market",
            "default": "global"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "competencies": {
            "items": {
              "$ref": "#/components/schemas/CompetencyCreate"
            },
            "type": "array",
            "title": "Competencies",
            "default": []
          },
          "created_by": {
            "type": "string",
            "maxLength": 100,
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "framework_id",
          "version",
          "created_by"
        ],
        "title": "CompetencyFrameworkCreate"
      },
      "CompetencyFrameworkListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompetencyFrameworkResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 100
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "CompetencyFrameworkListResponse"
      },
      "CompetencyFrameworkResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "framework_id": {
            "type": "string",
            "title": "Framework Id"
          },
          "domain_pack_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Pack Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "locale": {
            "type": "string",
            "title": "Locale"
          },
          "market": {
            "type": "string",
            "title": "Market"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "competencies": {
            "items": {
              "$ref": "#/components/schemas/CompetencyResponse"
            },
            "type": "array",
            "title": "Competencies",
            "default": []
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "framework_id",
          "version",
          "status",
          "locale",
          "market",
          "created_by"
        ],
        "title": "CompetencyFrameworkResponse"
      },
      "CompetencyFrameworkUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          }
        },
        "type": "object",
        "title": "CompetencyFrameworkUpdate"
      },
      "CompetencyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "competency_id": {
            "type": "string",
            "title": "Competency Id"
          },
          "framework_id": {
            "type": "string",
            "title": "Framework Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "cognitive_levels": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cognitive Levels"
          },
          "critical": {
            "type": "boolean",
            "title": "Critical"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "competency_id",
          "framework_id",
          "name",
          "critical",
          "weight",
          "sort_order"
        ],
        "title": "CompetencyResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}